+++ Leomard/AppStorage/UserPreferences.swift
@@ -44,6 +44,7 @@ final class UserPreferences: ObservableObject {
     @AppStorage("two_column_view", store: .standard) var twoColumnView: Bool = true
     @AppStorage("save_liked_posts", store: .standard) var saveLikedPosts: Bool = false
     @AppStorage("autosave_post_creation", store: .standard) var autosavePostCreation: Bool = true
+    @AppStorage("use_piped_video_for_youtube", store: .standard) var usePipedVideoForYoutube: Bool = false
     
     // MARK: - Navbar
     @AppStorage("navbar_show_letter_separators", store: .standard) var navbarShowLetterSeparators: Bool = false